[feat]회원탈퇴 예외반환

+++ server/src/main/java/ILearn/member/controller/MemberController.java
@@ -104,16 +104,31 @@ public class MemberController {
             @io.swagger.annotations.ApiResponse(code = 200, message = "탈퇴 성공"),
             @io.swagger.annotations.ApiResponse(code = 404, message = "유저를 찾을 수 없음"),
     })
+
+
     public ResponseEntity<ApiResponse<?>> deleteMember(@PathVariable Long user_id) {
         try {
-            memberService.deleteMember(user_id);
+            Member member = memberService.findVerifiedMember(user_id);
 
-            ApiResponse<Void> response = new ApiResponse<>(true, "success");
-            return ResponseEntity.ok(response);
+            if(member != null){
+                if(member.getMemberStatus() == Member.MemberStatus.MEMBER_ACTIVE) {
+                    memberService.deleteMember(user_id);
+                    ApiResponse<Void> response = new ApiResponse<>(true,0,"success",null);
+                    return ResponseEntity.ok(response);
+                } else {
 
-        } catch (ApiResponseException ex) {
+                    ApiResponse<Void> response = new ApiResponse<>(false,920,"USER_NOT_FOUND",null);
+                    return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
+                }
+            } else {
+                // 회원을 찾을 수 없는 경우에 대한 예외 처리
+                ApiResponse<Void> response = new ApiResponse<>(false, 920, "USER_NOT_FOUND", null);
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
+            }
+        } catch (ApiResponseException ex){
+
             ApiResponse<?> response = ex.getResponse();
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(response);
         }
     }
 }
